fixes osx builds - static_cast include missing headers

+++ src/wallet/WalletGreen.cpp
@@ -3958,10 +3958,10 @@ namespace CryptoNote
         {
             m_logger(ERROR, BRIGHT_RED) << "Fusion transaction threshold is too small. Threshold "
                                         << m_currency.formatAmount(threshold) << ", minimum threshold "
-                                        << m_currency.formatAmount(CryptoNote::parameters::FUSION_TX_MAX_POOL_AMOUNT_DUST_V1 + 1);
+                                        << m_currency.formatAmount(static_cast<uint64_t>(CryptoNote::parameters::FUSION_TX_MAX_POOL_AMOUNT_DUST_V1 + 1));
             throw std::runtime_error(
                 "Threshold must be greater than "
-                + m_currency.formatAmount(CryptoNote::parameters::FUSION_TX_MAX_POOL_AMOUNT_DUST_V1));
+                + m_currency.formatAmount(static_cast<uint64_t>(CryptoNote::parameters::FUSION_TX_MAX_POOL_AMOUNT_DUST_V1)));
         }
 
         if (m_walletsContainer.get<RandomAccessIndex>().size() == 0)
